// src/shared/api/models/getCategoriesTypeItem.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * MT Service API
 * REST API Documentation for MT Service
 * OpenAPI spec version: 1.0
 */

export type GetCategoriesTypeItem = typeof GetCategoriesTypeItem[keyof typeof GetCategoriesTypeItem];


export const GetCategoriesTypeItem = {
  INCOME: 'INCOME',
  EXPENSE: 'EXPENSE',
} as const;